Mehrzad / 67136 / refactor bot-footer-extentions to TS (#5701)
