+++ packages/bot-web-ui/src/components/bot-footer-extensions/index.ts
@@ -1,4 +1,4 @@
-import BotFooterExtensions from './bot-footer-extensions.jsx';
+import BotFooterExtensions from './bot-footer-extensions';
 import './bot-footer-extensions.scss';
 
 export default BotFooterExtensions;
