+++ packages/bot-web-ui/src/components/bot-footer-extensions/security-and-privacy.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Icon, Popover, StaticUrl } from '@deriv/components';
+import { localize } from '@deriv/translations';
+
+const SecurityAndPrivacy: React.FC = () => (
+    <Popover alignment='top' message={localize('Security and privacy')}>
+        <StaticUrl className='footer__link' href='tnc/security-and-privacy.pdf' is_document>
+            <Icon icon='IcSecurityAndPrivacy' />
+        </StaticUrl>
+    </Popover>
+);
+
+export default SecurityAndPrivacy;
