--- packages/bot-web-ui/src/constants/bot-contents.js
@@ -1,4 +1,0 @@
-export const tabs_title = Object.freeze({
-    WORKSPACE: 'Workspace',
-    CHART: 'Chart',
-});
