+++ packages/bot-web-ui/src/constants/bot-contents.ts
@@ -1,0 +1,8 @@
+type TTabsTitleProps = {
+    [key: string]: string;
+};
+
+export const tabs_title: TTabsTitleProps = Object.freeze({
+    WORKSPACE: 'Workspace',
+    CHART: 'Chart',
+});
